updateCart() test and implementation deployed, also ControllerTestUtilities class refactored
