+++ server/cart-service/src/main/java/com/fnd/games_store/cart/dto/CartRequestDTO.java
@@ -5,12 +5,14 @@ import com.fnd.games_store.cart.entity.Game;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 import java.util.Set;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@ToString
 public class CartRequestDTO {
 
     private String userId;
