+++ server/cart-service/src/test/java/com/fnd/games_store/cart/test/controller_integration_test/GetCartContentTest.java
@@ -6,9 +6,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fnd.games_store.cart.CartApplication;
 import com.fnd.games_store.cart.dto.CartRequestDTO;
 import com.fnd.games_store.cart.dto.CartResponseDTO;
+import com.fnd.games_store.cart.dto.GameResponseDTO;
 import com.fnd.games_store.cart.entity.Game;
 import com.fnd.games_store.cart.service.CartCrudService;
 import com.fnd.games_store.cart.test.utilities.ControllerTestUtilities;
+import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +28,7 @@ import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -33,19 +36,11 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest(classes = CartApplication.class)
 @AutoConfigureMockMvc
+@Slf4j
 public class GetCartContentTest extends ControllerTestUtilities {
 
-
-
-    @Autowired
-    private MockMvc mvc;
-
-    @MockBean
-    private CartCrudService service;
-
-
     @Test
-    void test_getCartContent() throws Exception {
+    void test_updateCartContent() throws Exception {
 
         MvcResult result = this.mvc.perform(post("/v1/update")
                                     .content(jsonCartRequestDTO(createAppropriateCartRequest(userId,testGameSet)))
@@ -54,6 +49,10 @@ public class GetCartContentTest extends ControllerTestUtilities {
                                     .andExpect(status().isOk())
                                     .andReturn();
 
+        String body = result.getResponse().getContentAsString();
+        CartResponseDTO response = new ObjectMapper().readValue(body, CartResponseDTO.class);
+
+        assertThat(response).isEqualTo(createAppropriateCartResponse(userId,testGameSet));
 
         verify(service).updateCart(createAppropriateCartRequest(userId,testGameSet));
     }
@@ -64,55 +63,11 @@ public class GetCartContentTest extends ControllerTestUtilities {
     void testSetup(){
         testGameSet.add(createTestGameEntity(1));
 
+        when(service.updateCart(createAppropriateCartRequest(userId,testGameSet))).thenReturn(createAppropriateCartResponse(userId, testGameSet));
+
     }
 
 
-    protected ObjectMapper objectMapper = new ObjectMapper();
-
-    protected String userId = "user_1";
-
-    protected Set<Game> testGameSet = new HashSet<>();
-
-
-
-
-    protected String jsonCartRequestDTO(CartRequestDTO cartRequestDTO) {
-
-        try {
-            return objectMapper.writeValueAsString(createAppropriateCartRequest(cartRequestDTO.getUserId(), cartRequestDTO.getGameData()));
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-            return "";
-        }
-    }
-
-
-
-    protected CartRequestDTO createAppropriateCartRequest(String userId, Set<Game> gameData) {
-            return new CartRequestDTO(userId, gameData);
-        }
-
-    protected CartResponseDTO createAppropriateCartResponse() {
-        return new CartResponseDTO(userId,testGameSet);
-    }
-
-
-    protected Game createTestGameEntity(int differenceParameter){
-        Game game = new Game();
-        game.setGameId("id"+ differenceParameter);
-        game.setName("name" + differenceParameter);
-        game.setGenre("genre" + differenceParameter);
-        game.setReleaseDate("date" + differenceParameter);
-        game.setDeveloper("developer" + differenceParameter);
-        game.setPublisher("publisher" + differenceParameter);
-        game.setPlatform("platform" + differenceParameter);
-        game.setFeatures("features" + differenceParameter);
-        game.setPrice(BigDecimal.valueOf(1000));
-        game.setDiscount(BigDecimal.valueOf(200));
-        game.setDescription("description" + differenceParameter);
-        game.setBase64Image("image" + differenceParameter);
-        return game;
-    }
 
 }
